requirements.txt
defusedxml==0.4.1

Django==1.10

django-crispy-forms==1.6.0

django-pagedown==0.1.1

django-widget-tweaks==1.4.1

djangorestframework==3.4.5

oauthlib==2.0.0

Pillow==3.3.1

PyJWT==1.4.2

python-social-auth==0.2.21

python3-openid==3.0.10

requests==2.11.1

requests-oauthlib==0.6.2

selenium==2.53.6

six==1.10.0
